requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==44.0.0
dj-database-url==2.3.0
Django==5.1.5
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
gunicorn==23.0.0
packaging==24.2
psycopg2==2.9.10
pycparser==2.22
python-dotenv==1.0.1
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
waitress==3.0.2
